# Optimize computation of gas fee