## Proposed changes

- Add the formula for better understanding of dynamic base fee calculation
- Do not calculate delta if the parent gas price already reached to the upper bound or the lower bound.
- The performance result:

### Keep touching the lower bound condition
    - Shortcut enabled
        - BenchmarkNextBlockBaseFeeLowerBound-10    	 6949518	       167.4 ns/op
    - Shortcut disabled
        - BenchmarkNextBlockBaseFeeLowerBound-10    	 4421775	       268.3 ns/op

**37% performance improved.**

### Keep touching the upper bound condition
    - Shortcut enabled
      - BenchmarkNextBlockBaseFeeUpperBound-10    	 7683355	       153.3 ns/op
    - Shortcut disabled
      - BenchmarkNextBlockBaseFeeUpperBound-10    	 4369462	       270.1 ns/op

**43% performance improved.**


### Random 
    - Shortcut enabled
      - BenchmarkNextBlockBaseFeeRandom-10    	 3875472	       315.2 ns/op
    - Shortcut disabled
      - BenchmarkNextBlockBaseFeeRandom-10    	 3912888	       302.1 ns/op

**4% performance degradation**

## Types of changes

Please put an x in the boxes related to your change.

- [ ] Bugfix
- [x] New feature or enhancement
- [ ] Others

## Checklist

*Put an x in the boxes that apply. You can also fill these out after creating the PR. If you're unsure about any of them, don't hesitate to ask. We're here to help! This is simply a reminder of what we are going to look for before merging your code.*

- [x] I have read the [CONTRIBUTING GUIDELINES](https://github.com/klaytn/klaytn/blob/master/CONTRIBUTING.md) doc
- [x] I have signed the [CLA](https://cla-assistant.io/klaytn/klaytn)
- [x] Lint and unit tests pass locally with my changes (`$ make test`)
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)
- [ ] Any dependent changes have been merged and published in downstream modules